backend/requirements.txt
beautifulsoup4==4.14.2
fastapi==0.121.1
langchain_core==1.0.4
langchain-community
langchain_google_genai==3.0.1
pydantic==2.12.4
python-dotenv==1.2.1
Requests==2.32.5
SQLAlchemy==2.0.44
uvicorn[standard]
mysql-connector-python
